Move get_secure_storage to helper in order to prevent circular import
